Fix issue #64: uncovertible images
When iText.text.pdf.parser fails to parse/convert an image, CERMINE
fails to catch the exception and crashes. This patch catches the
exception, reports it, and skips the page with the unconvertible image
(continue to parse the rest of the pages).

See issue #64: https://github.com/CeON/CERMINE/issues/64